Refactored a bit the test fixtures architecture
