+++ src/Domain/Packaging/Native/GenericDockerfile.php
@@ -1,0 +1,76 @@
+<?php declare(strict_types=1);
+
+namespace Kiboko\Cloud\Domain\Packaging\Native;
+
+use Kiboko\Cloud\Domain\Packaging;
+
+final class GenericDockerfile implements \IteratorAggregate, Packaging\PackageInterface, Packaging\BuildableInterface
+{
+    private Packaging\RepositoryInterface $repository;
+    private Packaging\Placeholder $path;
+
+    public function __construct(
+        Packaging\RepositoryInterface $repository,
+        Packaging\Placeholder $path
+    ) {
+        $this->repository = $repository;
+        $this->path = $path;
+    }
+
+    public function __invoke(): \Traversable
+    {
+        yield new Packaging\Context\BuildableContext(
+            null,
+            $this->path,
+            []
+        );
+    }
+
+    public function getIterator()
+    {
+        /** @var Packaging\Context\BuildableContextInterface $context */
+        foreach ($this() as $context) {
+            yield new GenericTag($this->repository, $context);
+        }
+    }
+
+    public function pull(Packaging\Execution\CommandBus\Task $task): Packaging\Execution\CommandBus\Task
+    {
+        /** @var Packaging\Tag\TagBuildInterface $tag */
+        foreach ($this as $tag) {
+            $tag->pull($task);
+        }
+
+        return $task;
+    }
+
+    public function push(Packaging\Execution\CommandBus\Task $task): Packaging\Execution\CommandBus\Task
+    {
+        /** @var Packaging\Tag\TagBuildInterface $tag */
+        foreach ($this as $tag) {
+            $tag->push($task);
+        }
+
+        return $task;
+    }
+
+    public function build(Packaging\Execution\CommandBus\Task $task): Packaging\Execution\CommandBus\Task
+    {
+        /** @var Packaging\Tag\TagBuildInterface $tag */
+        foreach ($this as $tag) {
+            $tag->build($task);
+        }
+
+        return $task;
+    }
+
+    public function forceBuild(Packaging\Execution\CommandBus\Task $task): Packaging\Execution\CommandBus\Task
+    {
+        /** @var Packaging\Tag\TagBuildInterface $tag */
+        foreach ($this as $tag) {
+            $tag->forceBuild($task);
+        }
+
+        return $task;
+    }
+}
