+++ test/Fixture/VisitableFixture.php
@@ -2,13 +2,11 @@
 
 namespace test\Kiboko\Cloud\Fixture;
 
-use test\Kiboko\Cloud\WizardAssertionFixtureProvider;
-
 trait VisitableFixture
 {
     public function apply(FixtureVisitorInterface ...$visitors): iterable
     {
-        /** @var WizardAssertionFixtureProvider $fixtureProvider */
+        /** @var FixtureProviderInterface $fixtureProvider */
         foreach ($this->get() as $fixtureProvider) {
             foreach ($visitors as $visitor) {
                 $fixtureProvider = $visitor($fixtureProvider);
