+++ test/Fixture/Visitor/WithBlackfire.php
@@ -3,11 +3,11 @@
 namespace test\Kiboko\Cloud\Fixture\Visitor;
 
 use test\Kiboko\Cloud\Fixture\FixtureVisitorInterface;
-use test\Kiboko\Cloud\WizardAssertionFixtureProvider;
+use test\Kiboko\Cloud\Fixture\FixtureProviderInterface;
 
 final class WithBlackfire implements FixtureVisitorInterface
 {
-    public function __invoke(WizardAssertionFixtureProvider $subject): WizardAssertionFixtureProvider
+    public function __invoke(FixtureProviderInterface $subject): FixtureProviderInterface
     {
         return $subject->withBlackfire();
     }
