+++ test/Fixture/Visitor/WithoutDockerForMacOptimizations.php
@@ -2,12 +2,12 @@
 
 namespace test\Kiboko\Cloud\Fixture\Visitor;
 
+use test\Kiboko\Cloud\Fixture\FixtureProviderInterface;
 use test\Kiboko\Cloud\Fixture\FixtureVisitorInterface;
-use test\Kiboko\Cloud\WizardAssertionFixtureProvider;
 
 final class WithoutDockerForMacOptimizations implements FixtureVisitorInterface
 {
-    public function __invoke(WizardAssertionFixtureProvider $subject): WizardAssertionFixtureProvider
+    public function __invoke(FixtureProviderInterface $subject): FixtureProviderInterface
     {
         return $subject->withoutDockerForMacOptimizations();
     }
